+++ CHANGES.md
@@ -25,6 +25,9 @@
 
 <!-- Changes to how Black is packaged, such as dependency requirements -->
 
+- Fix shutdown errors in PyInstaller builds on macOS by disabling multiprocessing in
+  frozen environments (#4930)
+
 ### Parser
 
 <!-- Changes to the parser or to version autodetection -->
@@ -1760,6 +1763,7 @@ and the first release covered by our new
 ## 18.9b0
 
 - numeric literals are now formatted by _Black_ (#452, #461, #464, #469):
+
   - numeric literals are normalized to include `_` separators on Python 3.6+ code
 
   - added `--skip-numeric-underscore-normalization` to disable the above behavior and
@@ -1809,6 +1813,7 @@ and the first release covered by our new
 - typing stub files (`.pyi`) now have blank lines added after constants (#340)
 
 - `# fmt: off` and `# fmt: on` are now much more dependable:
+
   - they now work also within bracket pairs (#329)
 
   - they now correctly work across function/class boundaries (#335)